fix: Search button not triggering backend
The search button needed to reconnect before submitting information to the backend. This is needed because socket io will be disconnected by a useEffect in the body any time the page is refreshed. Since I am using a single socket io instance, I need to reconnect to use that same instance once it is disconnected. Previously, however, I was recreating the instance everytime it was needed which didn't actually require me to reconnect on submit