fix: Socket connection not working except for rerender
I fixed this error by first sharing socket connection using a custom context globally instead of prop drilling. This is because redux doesn't work to store unserializable objects.
Then I was mannually udpating the socket object in a separate function stored in the buttonEventHandlers which didn't reflect the change on the object which was an issue too. I replace this function's instead of updating the socket object manually to change a state stored in redux store.
Next, I had to create a useEffect that will listen to the state change of the socket connection in the redux store and update the referenced object (the socket object assigned to a useRef object) connection accordingly after listening the socket connection.

This commit, however, deosn
't contain an efficient implementation yet, the socket connection isn't disconnected accross different requests made propery which I have to manually trigger somehow