fix: Socket connection not working after disconnection
The socket connection needed to remove the user from the session before disconnecting the user. However, it was trying to remove the user after he/she was already disconnecting which obviously won't work. But the nasty part is that this remove the user action will be queued even though it couldn't remove the user during this connection, and then as soon as a new connection is found, this action gets performed and removes the new user's connection which wasn't supposed to happen. Now that the frontend thinks that there is a new connection available, it tires to send the usuall connection. But this won't work because that cached remove user action has removed the new connection and the backend throughs an error